Release 1

+++ server/API/users/usersCtrl.js
@@ -26,18 +26,11 @@ function getUser(req, res) {
                 throw new Error("Couldn't load secret code from .env");
             const { userId } = req.cookies;
             if (!userId)
-                throw new Error("No authorized user"); //{
-            // if (!userId) {
-            //     res.status(401).send({ success: false, error: "No authorized user" });
-            //     return;
-            //   }
+                throw new Error("No authorized user !!!!!!!");
             const decodedUserId = jwt_simple_1.default.decode(userId, secret);
             const query = `SELECT * FROM \`movie-booking\`.\`users\` WHERE userID = '${decodedUserId.userID}'`;
-            //connection.query(query, async (error, results, fields) => {
-            // or
             database_1.default.query(query, [decodedUserId], (error, results) => {
                 if (error) {
-                    console.error("Error executing SQL query:", error);
                     res.status(500).send({ error: "Error executing SQL query" });
                 }
                 else {
@@ -59,35 +52,30 @@ function register(req, res) {
                 throw new Error("Not all fields are available from req.body");
             const { error } = userValidator_1.UserValidation.validate({ email, password });
             if (error) {
-                console.log(error);
-                return res.status(500).send({
-                    success: false,
-                    error: error.message,
-                });
+                return res.status(500).send({ success: false, error: "A user with this email address already exists." });
             }
             const salt = bcrypt_1.default.genSaltSync(saltRounds);
             const hash = bcrypt_1.default.hashSync(password, salt);
             const query = `INSERT INTO \`movie-booking\`.\`users\` (email, password) VALUES ("${email}", "${hash}");`;
             database_1.default.query(query, (error, results, fields) => {
                 if (error) {
-                    console.log(error);
                     return res.status(500).send({
                         success: false,
-                        error: "Failed to insert user data into database",
+                        error: "Failed to insert user data into database. Check your details. Perhaps you are trying to enter already registered data.",
                     });
                 }
                 const secret = process.env.JWT_SECRET;
                 if (!secret)
                     return res.status(500).send({ success: false, error: "Couldn't load secret code from .env" });
-                //@ts-ignore
-                const cookie = { userID: results.userID };
+                const insertId = results.insertId;
+                const cookie = { userID: insertId };
                 const JWTCookie = jwt_simple_1.default.encode(cookie, secret);
                 res.cookie("userId", JWTCookie);
                 res.send({ success: true, userArray: results });
             });
         }
         catch (error) {
-            res.status(500).send({ success: false, error });
+            res.status(500).send({ success: false, error: error.message });
         }
     });
 }
@@ -116,7 +104,6 @@ function login(req, res) {
                     res.send({ success: true, userArray: results });
                 }
                 catch (error) {
-                    //console.log(err);
                     res.status(500).send({ success: false, error: error.message });
                 }
             }));
@@ -127,6 +114,7 @@ function login(req, res) {
     });
 }
 exports.login = login;
+// 
 function updateUser(req, res) {
     return __awaiter(this, void 0, void 0, function* () {
         try {
@@ -136,7 +124,6 @@ function updateUser(req, res) {
             }
             const { error } = userValidator_1.UserValidation.validate({ email, password });
             if (error) {
-                console.log(error);
                 return res.status(500).send({
                     success: false,
                     error: error.message,
@@ -155,15 +142,14 @@ function updateUser(req, res) {
                 const secret = process.env.JWT_SECRET;
                 if (!secret)
                     return res.status(500).send({ success: false, error: "Couldn't load secret code from .env" });
-                //@ts-ignore
-                const cookie = { userID: results.userID };
+                const cookie = { userID: id };
                 const JWTCookie = jwt_simple_1.default.encode(cookie, secret);
                 res.cookie("userId", JWTCookie);
                 res.send({ success: true, userArray: results });
             });
         }
         catch (error) {
-            res.status(500).send({ success: false, error });
+            res.status(500).send({ success: false, error: error.message });
         }
     });
 }
@@ -172,20 +158,23 @@ function deleteUser(req, res) {
     return __awaiter(this, void 0, void 0, function* () {
         try {
             const id = req.params.id;
+            if (!id) {
+                return res.status(400).json({ error: "Missing user ID." });
+            }
+            res.clearCookie('userId');
             const query = `DELETE FROM \`movie-booking\`.\`users\` WHERE userID = ${id}`;
             database_1.default.query(query, (err, result) => {
                 if (err) {
-                    console.error("Error deleting user from MySQL database: ", err);
-                    return res.status(500).json({ error: "Something went wrong." });
+                    return res.status(500).json({ error: "Something went wrong. Error deleting user from the database." });
                 }
                 if (result.affectedRows === 0) {
-                    return res.status(500).json({ error: "No user found with the specified ID." });
+                    return res.status(404).json({ error: "No user found with the specified ID." });
                 }
                 return res.status(200).json({ success: "The user has been deleted." });
             });
         }
         catch (error) {
-            res.status(500).send({ error: error.message });
+            res.status(500).json({ success: false, error: error.message });
         }
     });
 }
@@ -197,7 +186,6 @@ function getUserByID(req, res) {
             const query = `SELECT * FROM  \`movie-booking\`.\`users\` WHERE userID = ${id}`;
             database_1.default.query(query, (err, result) => {
                 if (err) {
-                    console.error("Error getting user from MySQL database: ", err);
                     return res.status(500).json({ error: "Something went wrong." });
                 }
                 if (result.affectedRows === 0) {
@@ -207,57 +195,8 @@ function getUserByID(req, res) {
             });
         }
         catch (error) {
-            res.status(500).send({ error: error.message });
+            res.status(500).send({ success: false, error: error.message });
         }
     });
 }
 exports.getUserByID = getUserByID;
-// export async function getUser(req: express.Request, res: express.Response) {
-//   try {
-//     const secret: any = process.env.JWT_SECRET;
-//     if (!secret) throw new Error("Couldn't load secret code from .env");
-//     const userId = req.cookies; // Исправлено: req.cookies.userId вместо req.cookies
-//     console.log(userId);
-//     if (!userId) throw new Error("No authorized user");
-//     const decodedUserId = jwt.decode(userId, secret);
-//     console.log('---------decodedUserId----------');
-//     console.log(decodedUserId);
-//     const query = `SELECT * FROM \`movie-booking\`.\`users\` WHERE userID = '${decodedUserId.userID}'`;
-//     console.log(query);
-//     connection.query(query, (error, results) => {
-//       if (error) {
-//         console.error("Error executing SQL query:", error);
-//         res.status(500).send({ success: false, error: "Error executing SQL query" });
-//       } else {
-//         res.send({ success: true, userData: results });
-//       }
-//     });
-//   } catch (error: any) {
-//     res.status(500).send({ success: false, error: error.message });
-//   }
-// }
-// export async function getUser(req: express.Request, res: express.Response) {
-//     try {
-//         console.log('userId')
-//       const secret = process.env.JWT_SECRET;
-//       if (!secret) throw new Error("Couldn't load secret from .env");
-//       const { userID } = req.cookies;
-//       console.log(userID)
-//       if (!userID) throw new Error("Couldn't find user from cookies");
-//       const decodedUserId = jwt.decode(userID, secret);
-//       const { userId } = decodedUserId;  
-//       const query = `SELECT * FROM \`movie-booking\`.\`users\` WHERE userID = 105'`;
-//       connection.query(query, [decodedUserId], (error, results) => {
-//         if (error) {
-//             console.error("Error executing SQL query:", error);
-//             res.status(500).send({ error: "Error executing SQL query" });
-//         } else {
-//             // console.log("User data:", results);
-//             res.send({ sucess: true, userData: results });
-//             // res.send({ userDB });
-//         }
-//     });
-// } catch (error: any) {
-//     res.status(500).send({ sucess: false, error: error.message });
-// }
-// }
