+++ server/API/users/usersCtrl.ts
@@ -6,12 +6,6 @@ import connection from "../../DB/database";
 import { ResultSetHeader, RowDataPacket } from "mysql2";
 
 const saltRounds = 10;
-interface User {
-    userID: number;
-    email: string;
-    password: string;
-}
-
 
 export async function getUser(req: express.Request, res: express.Response) {
     try {
@@ -19,23 +13,12 @@ export async function getUser(req: express.Request, res: express.Response) {
 
         if (!secret) throw new Error("Couldn't load secret code from .env");
         const { userId } = req.cookies;
+        if (!userId) throw new Error("No authorized user !!!!!!!");
 
-       if (!userId) throw new Error("No authorized user");//{
-         
-        // if (!userId) {
-        //     res.status(401).send({ success: false, error: "No authorized user" });
-        //     return;
-        //   }
-       
         const decodedUserId = jwt.decode(userId, secret);
-
         const query = `SELECT * FROM \`movie-booking\`.\`users\` WHERE userID = '${decodedUserId.userID}'`;
-
-        //connection.query(query, async (error, results, fields) => {
-        // or
         connection.query(query, [decodedUserId], (error, results) => {
             if (error) {
-                console.error("Error executing SQL query:", error);
                 res.status(500).send({ error: "Error executing SQL query" });
             } else {
 
@@ -48,7 +31,6 @@ export async function getUser(req: express.Request, res: express.Response) {
     }
 }
 
-
 export async function register(req: express.Request, res: express.Response) {
     try {
         const { email, password } = req.body;
@@ -58,37 +40,36 @@ export async function register(req: express.Request, res: express.Response) {
 
         const { error } = UserValidation.validate({ email, password });
         if (error) {
-            console.log(error);
-            return res.status(500).send({
-                success: false,
-                error: error.message,
-            });
+            return res.status(500).send({ success: false, error: "A user with this email address already exists." });
         }
+
         const salt = bcrypt.genSaltSync(saltRounds);
         const hash = bcrypt.hashSync(password, salt);
 
         const query = `INSERT INTO \`movie-booking\`.\`users\` (email, password) VALUES ("${email}", "${hash}");`;
-        connection.query(query, (error, results, fields) => {
+        connection.query(query, (error, results: any, fields) => {
             if (error) {
-                console.log(error);
                 return res.status(500).send({
                     success: false,
-                    error: "Failed to insert user data into database",
+                    error: "Failed to insert user data into database. Check your details. Perhaps you are trying to enter already registered data.",
                 });
             }
 
             const secret = process.env.JWT_SECRET;
             if (!secret)
                 return res.status(500).send({ success: false, error: "Couldn't load secret code from .env" });
-            //@ts-ignore
-            const cookie = { userID: results.userID };
+
+            const insertId = results.insertId;
+
+            const cookie = { userID: insertId };
             const JWTCookie = jwt.encode(cookie, secret);
 
             res.cookie("userId", JWTCookie);
             res.send({ success: true, userArray: results });
         });
-    } catch (error) {
-        res.status(500).send({ success: false, error });
+
+    } catch (error: any) {
+        res.status(500).send({ success: false, error: error.message });
     }
 }
 
@@ -117,7 +98,6 @@ export async function login(req: express.Request, res: express.Response) {
                 res.cookie("userId", JWTCookie);
                 res.send({ success: true, userArray: results });
             } catch (error: any) {
-                //console.log(err);
                 res.status(500).send({ success: false, error: error.message });
             }
         });
@@ -125,6 +105,8 @@ export async function login(req: express.Request, res: express.Response) {
         res.status(500).send({ success: false, error: error.message });
     }
 }
+
+// 
 
 export async function updateUser(req: express.Request, res: express.Response) {
     try {
@@ -134,7 +116,6 @@ export async function updateUser(req: express.Request, res: express.Response) {
         }
         const { error } = UserValidation.validate({ email, password });
         if (error) {
-            console.log(error);
             return res.status(500).send({
                 success: false,
                 error: error.message,
@@ -156,37 +137,42 @@ export async function updateUser(req: express.Request, res: express.Response) {
             const secret = process.env.JWT_SECRET;
             if (!secret)
                 return res.status(500).send({ success: false, error: "Couldn't load secret code from .env" });
-            //@ts-ignore
-            const cookie = { userID: results.userID };
+
+            const cookie = { userID: id };
             const JWTCookie = jwt.encode(cookie, secret);
 
             res.cookie("userId", JWTCookie);
             res.send({ success: true, userArray: results });
         });
-    } catch (error) {
-        res.status(500).send({ success: false, error });
+    } catch (error: any) {
+        res.status(500).send({ success: false, error: error.message });
     }
 }
+
 
 export async function deleteUser(req: express.Request, res: express.Response) {
     try {
         const id = req.params.id;
+        if (!id) {
+            return res.status(400).json({ error: "Missing user ID." });
+        }
+
+        res.clearCookie('userId');
+
         const query = `DELETE FROM \`movie-booking\`.\`users\` WHERE userID = ${id}`;
         connection.query(query, (err, result: ResultSetHeader) => {
             if (err) {
-                console.error("Error deleting user from MySQL database: ", err);
-                return res.status(500).json({ error: "Something went wrong." });
+                return res.status(500).json({ error: "Something went wrong. Error deleting user from the database." });
             }
 
             if (result.affectedRows === 0) {
-                return res.status(500).json({ error: "No user found with the specified ID." });
+                return res.status(404).json({ error: "No user found with the specified ID." });
             }
 
             return res.status(200).json({ success: "The user has been deleted." });
-
         });
     } catch (error: any) {
-        res.status(500).send({ error: error.message });
+        res.status(500).json({ success: false, error: error.message });
     }
 }
 
@@ -197,7 +183,6 @@ export async function getUserByID(req: express.Request, res: express.Response) {
 
         connection.query(query, (err, result: ResultSetHeader) => {
             if (err) {
-                console.error("Error getting user from MySQL database: ", err);
                 return res.status(500).json({ error: "Something went wrong." });
             }
             if (result.affectedRows === 0) {
@@ -208,66 +193,8 @@ export async function getUserByID(req: express.Request, res: express.Response) {
 
         });
     } catch (error: any) {
-        res.status(500).send({ error: error.message });
+        res.status(500).send({ success: false, error: error.message });
     }
 }
 
-// export async function getUser(req: express.Request, res: express.Response) {
-//   try {
-//     const secret: any = process.env.JWT_SECRET;
-//     if (!secret) throw new Error("Couldn't load secret code from .env");
 
-//     const userId = req.cookies; // Исправлено: req.cookies.userId вместо req.cookies
-//     console.log(userId);
-//     if (!userId) throw new Error("No authorized user");
-
-//     const decodedUserId = jwt.decode(userId, secret);
-//     console.log('---------decodedUserId----------');
-//     console.log(decodedUserId);
-//     const query = `SELECT * FROM \`movie-booking\`.\`users\` WHERE userID = '${decodedUserId.userID}'`;
-//     console.log(query);
-
-//     connection.query(query, (error, results) => {
-//       if (error) {
-//         console.error("Error executing SQL query:", error);
-//         res.status(500).send({ success: false, error: "Error executing SQL query" });
-//       } else {
-//         res.send({ success: true, userData: results });
-//       }
-//     });
-//   } catch (error: any) {
-//     res.status(500).send({ success: false, error: error.message });
-//   }
-// }
-
-
-// export async function getUser(req: express.Request, res: express.Response) {
-//     try {
-//         console.log('userId')
-//       const secret = process.env.JWT_SECRET;
-//       if (!secret) throw new Error("Couldn't load secret from .env");
-
-//       const { userID } = req.cookies;
-//       console.log(userID)
-//       if (!userID) throw new Error("Couldn't find user from cookies");
-
-//       const decodedUserId = jwt.decode(userID, secret);
-//       const { userId } = decodedUserId;  
-
-
-//       const query = `SELECT * FROM \`movie-booking\`.\`users\` WHERE userID = 105'`;
-
-//       connection.query(query, [decodedUserId], (error, results) => {
-//         if (error) {
-//             console.error("Error executing SQL query:", error);
-//             res.status(500).send({ error: "Error executing SQL query" });
-//         } else {
-//             // console.log("User data:", results);
-//             res.send({ sucess: true, userData: results });
-//             // res.send({ userDB });
-//         }
-//     });
-// } catch (error: any) {
-//     res.status(500).send({ sucess: false, error: error.message });
-// }
-// }
